+++ jcop/tests/cz/cvut/felk/cig/jcop/problem/tsp/SwitchCityOperationTest.java
@@ -16,10 +16,10 @@ import org.testng.annotations.Test;
 public class SwitchCityOperationTest {
     @Test
     public void testExecute() throws Exception {
-        TSP problem = new TSP(new Double[][]{
-                {0.0, 1.0, 2.0},
-                {3.0, 0.0, 4.0},
-                {5.0, 6.0, 0.0},
+        TSP problem = new TSP(new Integer[][]{
+                {0, 1, 2},
+                {3, 0, 4},
+                {5, 6, 0},
         });
 
         Configuration startingConfiguration = problem.getStartingConfiguration();
